module_08/requirements.txt
alembic>=1.12
pika>=1.3
psycopg2-binary>=2.9
pydantic>=2.6
pymongo>=4.6
python-dotenv>=1.0
redis>=5.0
sqlalchemy>=2.0
motor>=3.3
